rename lib/tree.lua to lib/lib.lua
